add checks

+++ scripts/install_dev_soft_from_src.sh
@@ -11,20 +11,40 @@ if [ ! -x "$(command -v checkinstall)" ]; then
   echo Install tool for installation
   apt-get install -y \
     checkinstall
+
+  if [ $? -ne 0 ]; then
+    echo checkinstall can not be installed
+    exit 1
+  fi
 fi
 
 if [ ! -x "$(command -v automake)" ]; then
   echo Install automake
   apt-get install -y \
     automake
+
+  if [ $? -ne 0 ]; then
+    echo automake can not be installed
+    exit 1
+  fi
 fi
 
 if [ ! -x "$(command -v wget)" ]; then
   echo Install automake
   apt-get install -y \
     wget
+
+  if [ $? -ne 0 ]; then
+    echo wget can not be installed
+    exit 1
+  fi
 fi
 
+# add repository for 3.5 python (because bionic not contains the version)
+if [ "$(lsb_release -cs)" = "bionic" ]; then
+  add-apt-repository ppa:deadsnakes/ppa
+  apt-get update
+fi
 
 echo Prepare for install custom deb packages
 # For vim, vifm and vimb
@@ -36,7 +56,7 @@ apt-get install -y \
   libcairo2-dev \
   libatk1.0-dev \
   libx11-dev libxpm-dev libxt-dev \
-  liblua5.3 liblua5.3-dev \
+  liblua5.3-0 liblua5.3-dev \
   libperl-dev \
   ruby-dev \
   xz-utils libpthread-workqueue-dev \
@@ -45,6 +65,10 @@ apt-get install -y \
   python3.5-dev \
   libcurl4-openssl-dev
 
+if [ $? -ne 0 ]; then
+  echo neded packages can not be installed
+  exit 1
+fi
 
 echo Add links for lua
 LUA_DIR=/usr/include/lua5.3
@@ -65,6 +89,12 @@ echo --------------------------------------------------------------------------
 
 echo Install boost
 wget "https://dl.bintray.com/boostorg/release/1.69.0/source/boost_1_69_0.tar.gz"
+echo "9a2c2819310839ea373f42d69e733c339b4e9a19deab6bfec448281554aa4dbb boost_1_69_0.tar.gz" | sha256sum -c | grep -v OK
+if [ $? -eq 0 ]; then
+  echo boost can not be loaded
+  exit 1
+fi
+
 tar -xzvf boost_1_69_0.tar.gz
 rm boost_1_69_0.tar.gz
 cd boost_1_69_0
@@ -72,6 +102,10 @@ cd boost_1_69_0
 ./bootstrap.sh
 ./b2
 checkinstall -D -y --pkgname=boost-ch --pkgversion=1.69.0 --nodoc --backup=no ./b2 install
+if [ $? -ne 0 ]; then
+  echo boost can not be installed
+  exit 1
+fi
 
 cd ..
 rm -rf boost_1_69_0
@@ -80,6 +114,12 @@ echo --------------------------------------------------------------------------
 
 echo Install cmake
 wget "https://github.com/Kitware/CMake/releases/download/v3.14.1/cmake-3.14.1.tar.gz"
+echo "7321be640406338fc12590609c42b0fae7ea12980855c1be363d25dcd76bb25f  cmake-3.14.1.tar.gz" | sha256sum -c | grep -v OK
+if [ $? -eq 0 ]; then
+  echo cmake can not be loaded
+  exit 1
+fi
+
 tar -xzvf cmake-3.14.1.tar.gz
 rm cmake-3.14.1.tar.gz
 cd cmake-3.14.1
@@ -87,6 +127,10 @@ cd cmake-3.14.1
 ./bootstrap --system-curl
 make -j4
 checkinstall -D -y --pkgname=cmake-ch --pkgversion=3.14.1 --nodoc --backup=no
+if [ $? -ne 0 ]; then
+  echo cmake can not be installed
+  exit 1
+fi
 
 cd ..
 rm -rf cmake-3.14.1
@@ -105,6 +149,10 @@ cd ctags-master
 make -j4
 # real version of universal-ctags is 0.0.0, but we can not set it as version in checkinstall
 checkinstall -D -y --pkgname=ctags-ch --pkgversion=0.0.1 --nodoc --backup=no
+if [ $? -ne 0 ]; then
+  echo ctags can not be installed
+  exit 1
+fi
 
 cd ..
 rm -rf ctags-master
@@ -113,6 +161,12 @@ echo --------------------------------------------------------------------------
 
 echo Install vim
 wget "https://github.com/vim/vim/archive/v8.1.1140.tar.gz"
+echo "b2bd214f9e562308af7203e3e8cfeb13327d503ab2fe23090db9c42f13ca0145  v8.1.1140.tar.gz" | sha256sum -c | grep -v OK
+if [ $? -eq 0 ]; then
+  echo vim can not be loaded
+  exit 1
+fi
+
 tar -xzvf v8.1.1140.tar.gz
 rm v8.1.1140.tar.gz
 cd vim-8.1.1140
@@ -132,6 +186,10 @@ cd vim-8.1.1140
             --prefix=/usr/local
 make -j4
 checkinstall -D -y --pkgname=vim-ch --pkgversion=8.1.1140 --nodoc --backup=no
+if [ $? -ne 0 ]; then
+  echo vim can not be installed
+  exit 1
+fi
 
 cd ..
 rm -rf vim-8.1.1140
@@ -140,6 +198,12 @@ echo --------------------------------------------------------------------------
 
 echo Install vifm
 wget "https://github.com/vifm/vifm/archive/v0.10.tar.gz"
+echo "e5681c9e560e23d9deeec3b5b12e0ccad82612d9592c00407f3dd75cf5066548  v0.10.tar.gz" | sha256sum -c | grep -v OK
+if [ $? -eq 0 ]; then
+  echo vifm can not be loaded
+  exit 1
+fi
+
 tar -xzvf v0.10.tar.gz
 rm v0.10.tar.gz
 cd vifm-0.10
@@ -147,6 +211,10 @@ cd vifm-0.10
 ./configure
 make -j4
 checkinstall -D -y --pkgname=vifm-ch --pkgversion=0.10 --nodoc --backup=no
+if [ $? -ne 0 ]; then
+  echo vifm can not be installed
+  exit 1
+fi
 
 cd ..
 rm -rf vifm-0.10
@@ -155,12 +223,22 @@ echo --------------------------------------------------------------------------
 
 echo Install vimb
 wget "https://github.com/fanglingsu/vimb/archive/3.3.0.tar.gz"
+echo "5c6fe39b1b2ca18a342bb6683f7fd5b139ead53903f57dd9eecd5a1074576d6c  3.3.0.tar.gz" | sha256sum -c | grep -v OK
+if [ $? -eq 0 ]; then
+  echo vimb can not be loaded
+  exit 1
+fi
+
 tar -xzvf 3.3.0.tar.gz
 rm 3.3.0.tar.gz
 cd vimb-3.3.0
 
 make -j4
 checkinstall -D -y --pkgname=vimb-ch --pkgversion=3.3.0 --nodoc --backup=no
+if [ $? -ne 0 ]; then
+  echo vimb can not be installed
+  exit 1
+fi
 
 cd ..
 rm -rf vimb-3.3.0
